requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
alembic
python-dotenv
